feat: Bank 메서드 분리를 위한 ConsolManager 타입 구현
